+++ aether-datafixers-api/src/main/java/de/splatgames/aether/datafixers/api/dynamic/DynamicOps.java
@@ -25,6 +25,7 @@ package de.splatgames.aether.datafixers.api.dynamic;
 import com.google.common.base.Preconditions;
 import de.splatgames.aether.datafixers.api.result.DataResult;
 import de.splatgames.aether.datafixers.api.util.Pair;
+import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -237,6 +238,67 @@ public interface DynamicOps<T> {
      * @return the created value
      */
     @NotNull T createNumeric(@NotNull final Number value);
+
+    // ==================== Java Object Creation ====================
+
+    /**
+     * Creates a value from an arbitrary Java object. This is an optional,
+     * experimental method that may not be supported by all implementations.
+     *
+     * <p><b>Note:</b>
+     * Implementations of the {@code aether-datafixers-codec} module do not support
+     * this method and will throw {@link SecurityException} if called. This is because
+     * allowing arbitrary object creation can lead to security vulnerabilities if not
+     * used carefully. If you need to create values from Java objects, consider
+     * implementing a custom {@link de.splatgames.aether.datafixers.api.codec.Encoder Encoder}
+     * that explicitly defines how to convert your objects to the dynamic format.</p>
+     *
+     * @param value the Java object to convert
+     * @return the created value
+     * @throws UnsupportedOperationException if this method is not supported by the implementation
+     *         (default behavior)
+     * @throws SecurityException if the implementation explicitly rejects this operation
+     *         for security reasons (e.g., codec module implementations)
+     * @deprecated This is an experimental API that may not be implemented by all
+     *     DynamicOps. Prefer using typed creation methods (e.g., {@link #createString},
+     *     {@link #createInt}) or a custom Encoder for your specific use case.
+     * @since 1.1.0
+     */
+    @Deprecated
+    @ApiStatus.Experimental
+    default @NotNull T createObject(@NotNull final Object value) {
+        throw new UnsupportedOperationException("createObject is not supported by this DynamicOps implementation");
+    }
+
+    /**
+     * Reads an arbitrary Java object from a format value. This is the read counterpart
+     * to {@link #createObject(Object)} and is an optional, experimental method that may
+     * not be supported by all implementations.
+     *
+     * <p><b>Note:</b>
+     * Implementations of the {@code aether-datafixers-codec} module do not support
+     * this method and will throw {@link SecurityException} if called. This is because
+     * allowing arbitrary object deserialization can lead to security vulnerabilities
+     * if not used carefully. If you need to read values as Java objects, consider
+     * implementing a custom {@link de.splatgames.aether.datafixers.api.codec.Decoder Decoder}
+     * that explicitly defines how to convert the dynamic format to your objects.</p>
+     *
+     * @param input the format value to convert
+     * @return a {@link DataResult} containing the Java object on success, or an error
+     *         describing why the operation is not supported
+     * @throws SecurityException if the implementation explicitly rejects this operation
+     *         for security reasons (e.g., codec module implementations)
+     * @deprecated This is an experimental API that may not be implemented by all
+     *     DynamicOps. Prefer using typed reading methods (e.g., {@link #getStringValue},
+     *     {@link #getNumberValue}) or a custom Decoder for your specific use case.
+     * @see #createObject(Object)
+     * @since 1.1.0
+     */
+    @Deprecated
+    @ApiStatus.Experimental
+    default @NotNull DataResult<Object> getObjectValue(@NotNull final T input) {
+        return DataResult.error("getObjectValue is not supported by this DynamicOps implementation");
+    }
 
     // ==================== Primitive Reading ====================